feat(shared): provide mjs exports
