+++ packages/shared/src/config.ts
@@ -1,5 +1,5 @@
 import { join } from 'node:path';
-import fse from '../compiled/fs-extra';
+import fse from '../compiled/fs-extra/index.js';
 import { logger } from './logger';
 import type {
   InspectConfigOptions,
@@ -70,7 +70,9 @@ export async function outputInspectConfigFiles({
 }
 
 export async function stringifyConfig(config: unknown, verbose?: boolean) {
-  const { default: WebpackChain } = await import('../compiled/webpack-chain');
+  const { default: WebpackChain } = await import(
+    '../compiled/webpack-chain/index.js'
+  );
 
   // webpackChain.toString can be used as a common stringify method
   const stringify = WebpackChain.toString as (
