+++ packages/shared/src/logger.ts
@@ -1,5 +1,8 @@
-import { type Logger, logger } from '../compiled/rslog';
+import * as rslog from '../compiled/rslog/index.js';
+import type { Logger } from '../compiled/rslog/index.js';
 import { color } from './utils';
+
+const { logger } = rslog;
 
 // setup the logger level
 if (process.env.DEBUG) {
